Refactored code, added error handling for AuthenticationException thrown
