+++ ImageAnnotatorServer/src/main/java/com/vs/ImageAnnotatorServer/auth/AuthController.java
@@ -1,6 +1,6 @@
 package com.vs.ImageAnnotatorServer.auth;
 
-import com.vs.ImageAnnotatorServer.exception.UserCreationException;
+import com.vs.ImageAnnotatorServer.utils.exception.UserCreationException;
 import com.vs.ImageAnnotatorServer.rest.RestResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 
 
-// TODO Proper error handling, check for security smells and follow best practices
 @RestController
 @RequestMapping("/auth")
 public class AuthController {
