+++ ImageAnnotatorServer/src/main/java/com/vs/ImageAnnotatorServer/auth/AuthService.java
@@ -1,6 +1,6 @@
 package com.vs.ImageAnnotatorServer.auth;
 
-import com.vs.ImageAnnotatorServer.exception.UserCreationException;
+import com.vs.ImageAnnotatorServer.utils.exception.UserCreationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
